// crudapp2.0/src/app/model/product.ts
export class Product{
    movieId:number|any;
    movieName: string = '';
    movieDirector: string = '';
    movieGenre: string = '';
    movieReleaseDate: string = '';
    movieLanguage: string = '';
    duration: string = '';
    country: string = '';
    description: string = '';
    overallRate: number = 0;
    imageUrl:string|undefined;
    
    
}